refactor(wingman): extract CommandExecutor service
Move command/action execution out of the unified Wingman class into
a focused service. Covers get_command, instant activation matching,
command execution, and the keyboard/mouse/joystick action dispatcher.

Part of the post-merge cleanup to make wingman.py more navigable.
