+++ wingmen/wingman.py
@@ -53,6 +53,7 @@ from services.audio_library import AudioLibrary
 from services.conversation_manager import ConversationManager
 from services.conversation_condenser import ConversationCondenser
 from services.context_builder import ContextBuilder
+from services.command_executor import CommandExecutor
 from services.tool_executor import ToolExecutor
 from services.provider_factory import ProviderFactory
 from services.skill_registry import SkillRegistry
@@ -145,6 +146,13 @@ class Wingman:
         self.condenser = ConversationCondenser(self.conversation, config, name)
         self.context_builder = ContextBuilder(config, settings, name)
         self.tool_executor = ToolExecutor(config, settings, name)
+        self.command_executor = CommandExecutor(
+            config=config,
+            audio_library=audio_library,
+            wingman_name=name,
+            on_reset_history=self.reset_conversation_history,
+            on_add_forced_commands=self.add_forced_assistant_command_calls,
+        )
 
         # --- Token tracking ---
         self.last_turn_prompt_tokens: int = 0
@@ -934,7 +942,7 @@ class Wingman:
         await self.add_user_message(transcript, images=images)
 
         benchmark.start_snapshot("Instant activation commands")
-        instant_response, instant_command_executed = await self._try_instant_activation(
+        instant_response, instant_command_executed = await self.command_executor.try_instant_activation(
             transcript=transcript
         )
         if instant_response:
@@ -1151,7 +1159,7 @@ class Wingman:
 
         if response_message.tool_calls:
             response_message.tool_calls = await self.tool_executor.fix_tool_calls(
-                response_message.tool_calls, self.get_command
+                response_message.tool_calls, self.command_executor.get_command
             )
 
         prompt_tokens = 0
@@ -1176,8 +1184,8 @@ class Wingman:
             mcp_registry=self.mcp_registry,
             capability_registry=self.capability_registry,
             persistent_memory_service=self.persistent_memory_service,
-            get_command_fn=self.get_command,
-            execute_command_fn=self._execute_command,
+            get_command_fn=self.command_executor.get_command,
+            execute_command_fn=self.command_executor.execute_command,
             play_to_user_fn=self.play_to_user,
             local_ai_service=self.local_ai_service,
             update_tool_response_fn=self.conversation.update_tool_response,
@@ -1197,8 +1205,8 @@ class Wingman:
             mcp_registry=self.mcp_registry,
             capability_registry=self.capability_registry,
             persistent_memory_service=self.persistent_memory_service,
-            get_command_fn=self.get_command,
-            execute_command_fn=self._execute_command,
+            get_command_fn=self.command_executor.get_command,
+            execute_command_fn=self.command_executor.execute_command,
             play_to_user_fn=self.play_to_user,
         )
 
@@ -1399,29 +1407,6 @@ class Wingman:
                     traceback.format_exc(), color=LogType.ERROR, server_only=True
                 )
         return ""
-
-    # ───────────────── Instant activation ───────────────── #
-
-    async def _try_instant_activation(self, transcript: str) -> tuple[str, bool]:
-        commands = await self._execute_instant_activation_command(transcript)
-        if commands:
-            await self.add_forced_assistant_command_calls(commands)
-            responses = []
-            for command in commands:
-                if command.responses:
-                    responses.append(self._select_instant_command_response(command))
-
-            if len(responses) == len(commands):
-                responses = list(dict.fromkeys(responses))
-                responses = [
-                    response + "." if not response.endswith(".") else response
-                    for response in responses
-                ]
-                return " ".join(responses), True
-
-            return None, True
-
-        return None, False
 
     # ───────────────── Instant responses ───────────────── #
 
@@ -1711,192 +1696,15 @@ class Wingman:
 
         return tools
 
-    # ───────────────── Commands ─────────────────────────────── #
+    # ───────────────── Backward-compat delegation ────────────── #
 
     def get_command(self, command_name: str) -> CommandConfig | None:
-        if self.config.commands is None:
-            return None
-        command = next(
-            (item for item in self.config.commands if item.name == command_name),
-            None,
-        )
-        return command
-
-    def _select_instant_command_response(self, command: CommandConfig) -> str | None:
-        command_responses = command.responses
-        if (command_responses is None) or (len(command_responses) == 0):
-            return None
-        return random.choice(command_responses)
-
-    async def _execute_instant_activation_command(
-        self, transcript: str
-    ) -> list[CommandConfig] | None:
-        try:
-            commands_by_instant_activation = {}
-            for command in self.config.commands:
-                if command.instant_activation:
-                    for phrase in command.instant_activation:
-                        if phrase.lower() in commands_by_instant_activation:
-                            commands_by_instant_activation[phrase.lower()].append(
-                                command
-                            )
-                        else:
-                            commands_by_instant_activation[phrase.lower()] = [command]
-
-            phrase = difflib.get_close_matches(
-                transcript.lower(),
-                commands_by_instant_activation.keys(),
-                n=1,
-                cutoff=1,
-            )
-
-            if not phrase:
-                return None
-
-            commands = commands_by_instant_activation[phrase[0]]
-            for command in commands:
-                await self._execute_command(command, True)
-
-            return commands
-        except Exception as e:
-            await printr.print_async(
-                f"Error during instant activation in Wingman '{self.name}': {str(e)}",
-                color=LogType.ERROR,
-            )
-            printr.print(traceback.format_exc(), color=LogType.ERROR, server_only=True)
-            return None
-
-    async def _execute_command(
-        self, command: CommandConfig, is_instant=False
-    ) -> tuple[str | None, str]:
-        if not command:
-            return None, "Command not found"
-
-        try:
-            if len(command.actions or []) == 0:
-                await printr.print_async(
-                    f"No actions found for command: {command.name}",
-                    color=LogType.WARNING,
-                )
-            else:
-                await self.execute_action(command)
-                await printr.print_async(
-                    f"Executed {'instant' if is_instant else 'AI'} command: {command.name}",
-                    color=LogType.COMMAND,
-                )
-
-            if command.name == "ResetConversationHistory":
-                await self.reset_conversation_history()
-                await printr.print_async(
-                    f"Executed command: {command.name}", color=LogType.COMMAND
-                )
-
-            return (
-                self._select_instant_command_response(command),
-                command.additional_context or "OK",
-            )
-        except Exception as e:
-            await printr.print_async(
-                f"Error executing command '{command.name}' for Wingman '{self.name}': {str(e)}",
-                color=LogType.ERROR,
-            )
-            printr.print(traceback.format_exc(), color=LogType.ERROR, server_only=True)
-            return None, "ERROR DURING PROCESSING"
+        """Backward-compat: delegate to command_executor."""
+        return self.command_executor.get_command(command_name)
 
     async def execute_action(self, command: CommandConfig):
-        if not command or not command.actions:
-            return
-
-        def contains_numpad_key(hotkey: str) -> bool:
-            if not hotkey:
-                return False
-            tokens = hotkey.lower().split("+")
-            return any(token.startswith("num ") for token in tokens)
-
-        try:
-            for action in command.actions:
-                if action.keyboard:
-                    if action.keyboard.hotkey_codes and not contains_numpad_key(
-                        action.keyboard.hotkey
-                    ):
-                        code = action.keyboard.hotkey_codes
-                    else:
-                        code = action.keyboard.hotkey
-
-                    if action.keyboard.press == action.keyboard.release:
-                        hold = action.keyboard.hold or 0.1
-                        if (
-                            action.keyboard.hotkey_codes
-                            and len(action.keyboard.hotkey_codes) == 1
-                            and not contains_numpad_key(action.keyboard.hotkey)
-                        ):
-                            keyboard.direct_event(
-                                action.keyboard.hotkey_codes[0],
-                                0 + (1 if action.keyboard.hotkey_extended else 0),
-                            )
-                            time.sleep(hold)
-                            keyboard.direct_event(
-                                action.keyboard.hotkey_codes[0],
-                                2 + (1 if action.keyboard.hotkey_extended else 0),
-                            )
-                        else:
-                            keyboard.press(code)
-                            time.sleep(hold)
-                            keyboard.release(code)
-                    else:
-                        if (
-                            action.keyboard.hotkey_codes
-                            and len(action.keyboard.hotkey_codes) == 1
-                            and not contains_numpad_key(action.keyboard.hotkey)
-                        ):
-                            keyboard.direct_event(
-                                action.keyboard.hotkey_codes[0],
-                                (0 if action.keyboard.press else 2)
-                                + (1 if action.keyboard.hotkey_extended else 0),
-                            )
-                        else:
-                            keyboard.send(
-                                code,
-                                action.keyboard.press,
-                                action.keyboard.release,
-                            )
-
-                if action.mouse:
-                    if action.mouse.move_to:
-                        x, y = action.mouse.move_to
-                        mouse.move(x, y)
-
-                    if action.mouse.move:
-                        x, y = action.mouse.move
-                        mouse.move(x, y, absolute=False, duration=0.5)
-
-                    if action.mouse.scroll:
-                        mouse.wheel(action.mouse.scroll)
-
-                    if action.mouse.button:
-                        if action.mouse.hold:
-                            mouse.press(button=action.mouse.button)
-                            time.sleep(action.mouse.hold)
-                            mouse.release(button=action.mouse.button)
-                        else:
-                            mouse.click(button=action.mouse.button)
-
-                if action.write:
-                    keyboard.write(action.write)
-
-                if action.wait:
-                    time.sleep(action.wait)
-
-                if action.audio:
-                    await self.audio_library.handle_action(
-                        action.audio, self.config.sound.volume
-                    )
-        except Exception as e:
-            await printr.print_async(
-                f"Error executing actions of command '{command.name}' for wingman '{self.name}': {str(e)}",
-                color=LogType.ERROR,
-            )
-            printr.print(traceback.format_exc(), color=LogType.ERROR, server_only=True)
+        """Backward-compat: delegate to command_executor."""
+        await self.command_executor.execute_action(command)
 
     # ───────────────── Threading ─────────────────────────────── #
 
@@ -1934,6 +1742,7 @@ class Wingman:
                 old_config = deepcopy(self.config)
 
             self.config = config
+            self.command_executor.config = config
 
             await self._update_skill_configs(config)
